add: search cards
